Fix bug where it isn't possible to d&d a file when in orananalysis view
When in OranAnalysis view, it wasn't possible to drag & drop a new file in the main view, due to an exception being raised.

It seems to be better when only the visible box takes the binding for the drag & drop, so I assume it was related to some conflict between both treeview handling.